a: code spliting for analysis
